fix: resolve bulk upload spread error, modal close, toaster notification, credential modal and reload issue

+++ src/services/apiClient.ts
@@ -3,23 +3,25 @@ import axios from "axios";
 export const apiClient = axios.create({
   baseURL: import.meta.env.VITE_API_URL ?? "http://localhost:5000",
   headers: {
-    'Content-Type': 'application/json',
+    "Content-Type": "application/json",
   },
 });
-const PUBLIC_URLS = ["/users"]
+const PUBLIC_URLS = ["/users"];
 apiClient.interceptors.request.use((config) => {
-  
-  if (PUBLIC_URLS.some((url)=> config.url?.includes(url))) return config;
+  if (PUBLIC_URLS.some((url) => config.url?.includes(url))) return config;
 
   const expiry = localStorage.getItem("expiry");
 
   if (expiry && Date.now() > Number(expiry)) {
-    console.log("Token expired");
 
-    localStorage.clear();
-    window.dispatchEvent(new Event("auth:expired"))
+    if (!window.__AUTH_EXPIRED__) {
+      window.__AUTH_EXPIRED__ = true;
 
-    return Promise.reject(new Error("Token expired"));
+      localStorage.clear();
+      window.dispatchEvent(new Event("auth:expired"));
+    }
+
+     return Promise.reject(new Error("Token expired"));
   }
 
   const token = localStorage.getItem("accessToken");
@@ -36,10 +38,14 @@ apiClient.interceptors.response.use(
     if (error.response?.status === 401) {
       console.log("Unauthorized");
 
-      localStorage.clear();
-      window.dispatchEvent(new Event("auth:expired"))
+      if (!window.__AUTH_EXPIRED__) {
+        window.__AUTH_EXPIRED__ = true;
+
+        localStorage.clear();
+        window.dispatchEvent(new Event("auth:expired"));
+      }
     }
 
     return Promise.reject(error);
-  }
+  },
 );
